% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ok_proj6.R
\name{ok_proj6}
\alias{ok_proj6}
\title{Is 'PROJ library >= 6' available}
\usage{
ok_proj6()
}
\value{
logical, \code{TRUE} if the system library 'PROJ >= 6'
}
\description{
Test for availability of 'PROJ' system library version 6 or higher.
}
\details{
On unix-alikes, this function is run in \code{.onLoad()} to check that version 6 functionality is
available. On Windows, the load process sets the data file location with the version 6 API, and that
is used as a test instead.

If 'PROJ' library version 6 is not available, the package still compiles and installs
but is not functional.

The lack of function can be simulated by setting
\code{options(reproj.mock.noproj6 = TRUE)}, designed for use with the reproj package.
}
\examples{
ok_proj6()
}
