% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_create.R
\name{proj_create}
\alias{proj_create}
\title{Generate a projection string.}
\usage{
proj_create(source, format = 0L)
}
\arguments{
\item{source}{input projection specification (PROJ4, WKT2, PROJJSON, ...)}

\item{format}{integer, 0 for WKT, 1 for PROJ, 2 for PROJJSON}
}
\value{
character string in requested format
}
\description{
Input any accepted format of PROJ crs specification. Return value is a string
in the requested format.
}
\details{
See the \href{https://proj.org/development/reference/functions.html#transformation-setup}{library documentation}
for details on input and output formats.

Some nuances of the format are not available, currently we use formats
'0: PJ_WKT2_2018', '1: PJ_PROJ_5', '2: PROJJSON'.
}
\examples{
if (ok_proj6()) {
proj_create("EPSG:4326", format = 1)

proj_create("urn:ogc:def:crs:EPSG::4326")

proj_create("urn:ogc:def:crs:EPSG::4326", format = 2)

cat(wkt <- proj_create("EPSG:3857"))
proj_create(wkt, format = 1L)

wkt_method <- proj_create("+proj=etmerc +lat_0=38 +lon_0=125 +ellps=bessel")

cat(wkt_method)
proj_create(wkt, format = 1L)
}
}
