% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_trans.R
\name{proj_trans}
\alias{proj_trans}
\title{Transform a set of coordinates with 'PROJ'}
\usage{
proj_trans(x, target_crs, source_crs = NULL, ..., use_z = NA, use_m = NA)
}
\arguments{
\item{x}{input coordinates (x,y, list or matrix see \code{z_} and \code{t_})}

\item{source_crs, target_crs}{Source/Target CRS definition, coerced with \code{\link[wk:wk_crs_proj_definition]{wk::wk_crs_proj_definition()}}}

\item{...}{Additional parameters forwarded to \code{\link[wk:wk_handle]{wk::wk_handle()}}}

\item{use_z, use_m}{Used to declare the output type. Use \code{TRUE} to
ensure the output has that dimension, \code{FALSE} to ensure it does not,
and \code{NA} to leave the dimension unchanged.}
}
\value{
list of transformed coordinates, with 4- or 2-elements \code{x_}, \code{y_}, \code{z_}, \code{t_}
}
\description{
A raw interface to 'proj_trans' in 'PROJ => 6', if it is available.
}
\details{
Input 'x' is assumed to be 2-columns of "x", then "y" coordinates. If "z" or
"t" is required pass these in as named vectors with "z_" and "t_". For simplifying reasons
\code{z_} and \code{t_} must always match the length of \code{x} \code{y}. Both default to 0, and are automatically
recycled to the number of rows in \code{x}.

Values that are detected out of bounds by library PROJ are allowed, we return \code{Inf} in this
case, rather than the error "tolerance condition error".
}
\examples{
proj_trans(cbind(147, -42), "+proj=laea type=crs", "OGC:CRS84")
proj_trans(cbind(147, -42, -2), "+proj=laea type=crs", "OGC:CRS84")
proj_trans(cbind(147, -42, -2, 1), "+proj=laea type=crs", "OGC:CRS84")
proj_trans(wk::xy(147, -42, crs = "OGC:CRS84"), "+proj=laea type=crs")
proj_trans(wk::wkt("POLYGON ((1 1, 0 1, 0 0, 1 0, 1 1))", crs = "OGC:CRS84"), 3112)

}
\references{
see the \href{https://proj.org/development/reference/functions.html#coordinate-transformation}{PROJ library documentation}
for details on the underlying functionality
}
