% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/project.R
\name{project}
\alias{project}
\title{Project points}
\usage{
project(pts, from = NULL, to)
}
\arguments{
\item{pts}{A vector or list of points.}

\item{from,to}{The from and to projection. A character string. If from is left NULL, then it is
assumed to be projected from WGS84.}
}
\description{
Project points
}
\details{
New lines are stripped out of projection strings to facilitate collapsing them
onto multiple lines.
}
\examples{
pts <- c(-71,41)
to = '"+proj=gnom +lat_0=90 +lon_0=0 +x_0=6300000 +y_0=6300000 +ellps=WGS84
   +datum=WGS84 +units=m +no_defs"'
project(pts, to = to)
}

