% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproj_PROJ.R
\name{reproj_PROJ}
\alias{reproj_PROJ}
\title{Reproj base function}
\usage{
reproj_PROJ(x, target, ..., source = NULL)
}
\arguments{
\item{x}{matrix of x, y, z (z is optional, 2 or 3 columns)}

\item{target}{projection to transform to}

\item{...}{reserved}

\item{source}{projection being transformed from}
}
\value{
matrix of 3 columns (x, y, z or long, lat, z)
}
\description{
Coordinate transformation via the PROJ library
}
\details{
Designed for use by the reproj package.
}
\examples{
m <- cbind(w$x, w$y)
plot(m, pch = ".")
s0 <- "+init=epsg:4326"
t1 <- "+proj=laea"
m <- rePROJ(m, t1, source = s0); s0 <- t1
plot(m, pch = ".")
t1 <- "+proj=lcc +lon_0=147"
m <- rePROJ(m, t1, source = s0); s0 <- t1
plot(m, pch = ".")
t1 <- "+proj=gnom +lon_0=147"
m <- rePROJ(m, t1, source = s0); s0 <- t1
plot(m, pch = ".")
t1 <- "+proj=laea +lon_0=147"
m <- rePROJ(m, t1, source = s0); s0 <- t1
plot(m, pch = ".")
t1 <- "+proj=ortho +lon_0=147"
m <- rePROJ(m, t1, source = s0); s0 <- t1
plot(m, pch = ".")
t1 <- "+proj=merc +lon_0=147"
m <- rePROJ(m, t1, source = s0); s0 <- t1
plot(m, pch = ".")
}
