% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROJ-package.R
\docType{data}
\name{xymap}
\alias{xymap}
\title{xymap data for testing}
\description{
A copy of the xymap data set from the quadmesh package.
}
\details{
A matrix of longitude/latitude values of the world coastline.
}
