% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROJ-package.R
\docType{package}
\name{PROJ-package}
\alias{PROJ}
\alias{PROJ-package}
\title{PROJ: Generic Coordinate System Transformations}
\description{
'PROJ' is a generic coordinate transformation software that transforms 
 geospatial coordinates from one coordinate reference system (CRS) to another. This 
  includes cartographic projections as well as geodetic transformations.
}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com} (0000-0002-2471-7511)

}
\keyword{internal}
