% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROJ-package.R
\docType{package}
\name{PROJ-package}
\alias{PROJ}
\alias{PROJ-package}
\title{PROJ: Generic Coordinate System Transformations Using 'PROJ'}
\description{
'PROJ' is a generic coordinate transformation software that transforms 
 geospatial coordinates from one coordinate reference system (CRS) to another. This 
  includes cartographic projections as well as geodetic transformations. Version 5.0.0 or 
  higher is required.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/PROJ}
  \item Report bugs at \url{https://github.com/hypertidy/PROJ/issues}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com} (0000-0002-2471-7511)

}
\keyword{internal}
