% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROJ-package.R
\docType{package}
\name{PROJ-package}
\alias{PROJ}
\alias{PROJ-package}
\title{PROJ: Generic Coordinate System Transformations Using 'PROJ'}
\description{
Currently non-operational, a harmless wrapper to allow package 'reproj' to install and 
  function while relying on the 'proj4' package.

The goal of PROJ is to provide generic coordinate system transformations in R
without requiring bespoke formats for storing basic data.
}
\section{I. Functions}{

\tabular{ll}{
\code{\link{ok_proj6}} determine if PROJ version >=6 is available
\code{\link{proj_trans_generic}} a light wrapper around the underlying transformation functionality of PROJ version 6 (or higher)
removed   proj_create convert projection strings to WKT2 or PROJ format
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/PROJ}
  \item Report bugs at \url{https://github.com/hypertidy/PROJ/issues}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com} (\href{https://orcid.org/0000-0002-2471-7511}{ORCID})

Other contributors:
\itemize{
  \item Jeroen Ooms (provided PROJ library support on Windows, and assistance with Windows configuration) [contributor]
  \item Simon Urbanek (wrote original code versions for PROJ version 6) [copyright holder, contributor]
  \item Dewey Dunnington (key code contributions, and provided libproj to improve things a lot) [contributor]
}

}
\keyword{internal}
