% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wk-trans.R
\name{proj_create}
\alias{proj_create}
\title{Create PROJ transformation object}
\usage{
proj_create(source_crs, target_crs, use_z = NA, use_m = NA)
}
\arguments{
\item{source_crs, target_crs}{Source/Target CRS definition, coerced with \code{\link[wk:wk_crs_proj_definition]{wk::wk_crs_proj_definition()}}}

\item{use_z, use_m}{Used to declare the output type. Use \code{TRUE} to
ensure the output has that dimension, \code{FALSE} to ensure it does not,
and \code{NA} to leave the dimension unchanged.}
}
\value{
A PROJ transformation object
}
\description{
Creates a PROJ CRS to CRS transformation object, to be used in \code{\link[wk:wk_transform]{wk::wk_transform()}}.
}
