% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{proj_trans}
\alias{proj_trans}
\title{PROJ trans}
\usage{
proj_trans(TARGET, X, Y, Z, INV)
}
\arguments{
\item{TARGET}{target projection}

\item{X}{x coordinate}

\item{Y}{y coordinate}

\item{Z}{z coordinate}

\item{INV}{forward or inverse projection (default forward = FALSE)}
}
\value{
list of transformed x, y, z
}
\description{
Coordinate transform in forward or inverse mode
}
\details{
t is set to 0 internally
https://proj4.org/development/quickstart.html
}
\examples{
dst<- "+proj=laea +datum=WGS84 +lon_0=1"
proj_trans(dst, 0, 0, 0, FALSE)
proj_trans(dst, -111318.1, 0, 0, TRUE)
}
