% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_trans.R
\name{proj_trans}
\alias{proj_trans}
\title{PROJ trans}
\usage{
proj_trans(TARGET, X, Y, Z = 0, ..., INV = FALSE, quiet = FALSE)
}
\arguments{
\item{TARGET}{target projection}

\item{X}{x coordinate, no default}

\item{Y}{y coordinate, no default}

\item{Z}{z coordinate, defaults to zero}

\item{...}{unused, but ensures that 'INV' must be named}

\item{INV}{forward or inverse projection (default forward = FALSE)}

\item{quiet}{logical, emit messages and warnings, set to \code{TRUE} for not}
}
\value{
list of transformed x, y, z
}
\description{
Coordinate transform in forward or inverse mode
}
\details{
t is set to 0 internally
https://proj4.org/development/quickstart.html

The minimal input is three arguments, 'TARGET', 'X', 'Y'
}
\examples{
dst<- "+proj=laea +datum=WGS84 +lon_0=1"
proj_trans(dst, 0, 0)
proj_trans(dst, -111318.1, 0, INV = TRUE)
}
