% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_trans_generic.R
\name{proj_trans_generic}
\alias{proj_trans_generic}
\title{PROJ transform generic}
\usage{
proj_trans_generic(x, target, ..., source = NULL)
}
\arguments{
\item{x}{input coordinates}

\item{target}{projection for output coordinates}

\item{source}{projection of input coordinates (must be named)}
}
\description{
A raw interface to proj_trans_generic in PROJ => 6, if it is available.
}
\examples{
proj_trans_generic(cbind(147, -42), "+proj=laea", source = "epsg:4326")
}
