% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_version.R
\name{proj_version}
\alias{proj_version}
\title{Report PROJ library version}
\usage{
proj_version()
}
\value{
character string (major.minor.patch)
}
\description{
This function returns NA if PROJ lib is not available.
}
\examples{
proj_version()
}
